+++ Utilities/utilities.cpp
@@ -12,13 +12,17 @@ QString Utilities::formatMemory(qulonglong bytes, int precision)
     qulonglong x = 1024;
     if (bytes == 0) return "0";
     if (bytes < x) return QString::number(bytes) + " bytes";
-    if (bytes < x * 1024) return QString::number((float)bytes / x, 'f', precision) + " KB";
+    if (bytes < x * 1024)
+        return QString::number(static_cast<double>(bytes) / x, 'f', precision) + " KB";
     x *= 1024;
-    if (bytes < (x * 1024)) return QString::number((float)bytes / x, 'f', precision) + " MB";
+    if (bytes < (x * 1024))
+        return QString::number(static_cast<double>(bytes) / x, 'f', precision) + " MB";
     x *= 1024;
-    if (bytes < (x * 1024)) return QString::number((float)bytes / x, 'f', precision) + " GB";
+    if (bytes < (x * 1024))
+        return QString::number(static_cast<double>(bytes) / x, 'f', precision) + " GB";
     x *= 1024;
-    if (bytes < (x * 1024)) return QString::number((float)bytes / x, 'f', precision) + " TB";
+    if (bytes < (x * 1024))
+        return QString::number(static_cast<double>(bytes) / x, 'f', precision) + " TB";
     return "More than TB";
 }
